qr_code  수정

+++ contents/admin.py
@@ -1,7 +1,8 @@
 from django.contrib import admin
 from django import forms
-import os
 from django.utils.safestring import mark_safe
+from django.core.exceptions import ValidationError
+import os
 from .models import *
 
 class LikeInline(admin.TabularInline):
@@ -24,6 +25,17 @@ class ContentsPasswordInline(admin.TabularInline):
 
 class AddQRDatasForm(forms.ModelForm):
 
+    def clean_images(self):
+        images = self.files.getlist('images')
+        for image in images:
+            if image.name.find('arz.kr_') == -1:
+                raise forms.ValidationError('이미지명 형식이 올바르지 않습니다. 다시 등록해주세요.')
+            try:
+                name_slice = image.name.split('_')
+                qr_data = name_slice[0] + '_' + name_slice[1]
+                contents_type = name_slice[1][3:4]
+            except:
+                raise forms.ValidationError('이미지명 형식이 올바르지 않습니다. 다시 등록해주세요.')
     class Meta:
         model = QRDatas
         exclude = ['qr_data', 'is_active', 'contents_type', 'activation_code', 'login_admin']
@@ -35,15 +47,16 @@ class ChangeQRDatasForm(forms.ModelForm):
         model = QRDatas
         exclude = ['images']
 
-
+from django.http import HttpResponseNotFound
 @admin.register(QRDatas)
 class QRDatasAdmin(admin.ModelAdmin):
 
     def delete_queryset(self, request, queryset):
         for obj in queryset:
-          if os.path.isfile(obj.images.path):
-            os.remove(obj.images.path)
-            obj.delete()
+          if obj.images and hasattr(obj.images, 'url'):
+                if os.path.isfile(obj.images.path):
+                    os.remove(obj.images.path)
+          obj.delete()
 
     def delete_model(self, request, obj):
         os.remove(obj.images.path)
@@ -51,7 +64,7 @@ class QRDatasAdmin(admin.ModelAdmin):
 
     list_display = ('qr_data', 'is_active', 'contents_type', 'activation_code', 'contents_title', 'username', 'create_dt')
     list_display_links = ['qr_data']
-    list_editable = ['is_active', 'activation_code']
+    list_editable = ['is_active',]
     ordering = ('-qr_data',)
     # 필터링 항목 설정
     list_filter = ('is_active', 'contents_type',)
@@ -75,7 +88,7 @@ class QRDatasAdmin(admin.ModelAdmin):
             self.form = AddQRDatasForm
         else:
             self.form = self.change_form
-            self.readonly_fields = ('qr_data', 'contents_type','login_admin', 'create_dt','qrcode_image')
+            self.readonly_fields = ('qr_data', 'contents_type', 'activation_code', 'login_admin', 'create_dt', 'qrcode_image', )
 
         return super(QRDatasAdmin, self).get_form(request, obj, **kwargs)
 
@@ -86,10 +99,11 @@ class QRDatasAdmin(admin.ModelAdmin):
             name_slice = f.name.split('_')
             qr_data = name_slice[0] + '_' + name_slice[1]
             contents_type = name_slice[1][3:4]
+            activation_code = name_slice[2].split('.')
             try:
                 QRDatas.objects.get(qr_data=qr_data)
             except:
-                QRDatas.objects.create(images=f, qr_data=qr_data, contents_type=contents_type, login_admin=request.user)
+               QRDatas.objects.create(images=f, qr_data=qr_data, contents_type=contents_type, activation_code=activation_code[0], login_admin=request.user)
 
 @admin.register(Contents)
 class ContentsAdmin(admin.ModelAdmin):
